Fix: corrections to reflect merge

+++ IFX_BUILD.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 set -e
 source ../SETTINGS_AND_VERSIONS.sh
-	
+
+
+# Monolothic needed to produce isntallable tf_tfl_translate	
 BAZEL_CXX_BUILD_SETTINGS=(
 		--config=opt
+                --config=monolithic
 )
 BAZEL_REPO_OVERRIDES=(  )
 TARGET_ARCH=native
@@ -26,15 +29,23 @@ do
     "--no-tflite")
         NOTFLITE=1
         ;;
+    "--no-install")
+        NOINSTALL=1
+        ;;
     "--fast")
 	BAZEL_CXX_BUILD_SETTINGS=(
                   --copt=-O1 --cxxopt=-O1 --strip=never
+                  --config=monolithic
 	)
 	;;
     "--debug")
+	# includes workaround for mis-documented and buggy 
+	# per_object_debug_info feature.
 	BAZEL_CXX_BUILD_SETTINGS=(
-                --config=dbg
-                --copt=-gsplit-dwarf --copt=-O1 --cxxopt=-gsplit-dwarf --cxxopt=-O1 
+                --config=monolithic --config=dbg
+                --features=per_object_debug_info
+                --define='per_object_debug_info_file=yes'
+                --copt=-O0 --cxxopt=-O0 
                 --strip=never --fission=yes 
 	)
 	;;
@@ -94,7 +105,6 @@ then
       --local_cpu_resources="$LOCALJOBS" "${VERBOSE[@]}"
       "${BAZEL_REMOTE_OPTIONS[@]}"
       "${BAZEL_CXX_BUILD_SETTINGS[@]}"
-      --config=monolithic
      --cxxopt=-DTF_LITE_DISABLE_X86_NEON --copt=-DTF_LITE_DISABLE_X86_NEON
 	    "--repository_cache=/home/aifordes.work/bazel_repo_cache" 
 	    # Note: Cannot enable debug non-NFS scratch for bazel_root too small...
@@ -107,7 +117,6 @@ else
       --local_cpu_resources="$LOCALJOBS"  "${VERBOSE[@]}"
       "${BAZEL_REMOTE_OPTIONS[@]}"
       "${BAZEL_CXX_BUILD_SETTINGS[@]}"
-      --config=monolithic
       --cxxopt=-DTF_LITE_DISABLE_X86_NEON --copt=-DTF_LITE_DISABLE_X86_NEON
       --verbose_failures=yes
   )
@@ -171,7 +180,7 @@ else
   echo Skipping configure...
 fi
 
-BAZEL_TARGETS=( //tensorflow/lite/toco:toco //tensorflow/compiler/mlir/lite:tf_tfl_translate )
+BAZEL_TARGETS=( //tensorflow/compiler/mlir/lite:tf_tfl_translate )
 
 if [ -z "$NOBUILD" ]
 then
@@ -180,7 +189,6 @@ then
     #bazel build "${BAZEL_OPTIONS[@]}" //third_party/aws:aws || true  # EXPECTED FAILURE but needed to unpack packages
     #bazel build --local_cpu_resources="$LOCALJOBS" --config=dbg --strip=never "${VERBOSE[@]}" //tensorflow/compiler/mlir/lite:tf_tfl_translate
     #sed -e'1,$s/"+[cd]/"+g/g' -i bazel-$(basename $(pwd))/external/aws-checksums/source/intel/crc32c_sse42_asm.c  #  BUILD FAILS MISERABG:Y WITH GCC7 FFS
-  (
     # Wild patching orgy is needed... including yes in 2020
     # workarounds for pathname/command-line length limitations.
     if [ -n "$MINGW64_HOST" ] 
@@ -190,12 +198,15 @@ then
     # 
     echo bazel build "${BAZEL_CMDLINE_OPTIONS[@]}"  "${BAZEL_TARGETS[@]}"
     bazel build   "${BAZEL_CMDLINE_OPTIONS[@]}"  "${BAZEL_TARGETS[@]}"
-    mkdir -p ${TFLITE_MICRO_ROOT}/bin  
-    rm -f ${TFLITE_MICRO_ROOT}/bin/*
-    cp bazel-bin/tensorflow/compiler/mlir/lite/tf_tfl_translate${EXE_SUFFIX} \
+    if [ -z "$NOINSTALL" ]
+    then
+      mkdir -p ${TFLITE_MICRO_ROOT}/bin  
+      rm -f ${TFLITE_MICRO_ROOT}/bin/*
+      echo Installing to ${TFLITE_MICRO_ROOT}/bin
+      cp bazel-bin/tensorflow/compiler/mlir/lite/tf_tfl_translate${EXE_SUFFIX} \
         bazel-bin/tensorflow/lite/toco/toco${EXE_SUFFIX} \
         ${TFLITE_MICRO_ROOT}/bin
-  )
+    fi
 fi
 
 if [ -z "$NOTFLITE" ] 
@@ -210,10 +221,14 @@ then
     echo make -j 4 BUILD_TYPE=debug test_executables
     make -j 4 BUILD_TYPE=debug microlite
 
-    # Actual payload - installed confiured copy of tflite(u) library and makefiles
-    make TARGET=ifx_riscv32_install_only ${RISCV_SETTINGS[@]} install
-    # Set TAGS to use the portable_optimized kernels (by default) instead of the reference ones.
-    echo 'TAGS := portable_optimized' >> ${TFLITE_MICRO_ROOT}/tools/make/installed_settings.inc
+
+    if [ -z "$NOINSTALL" ]
+    then
+        # Actual payload - installed confiured copy of tflite(u) library and makefiles
+        make TARGET=ifx_riscv32_install_only ${RISCV_SETTINGS[@]} install
+        # Set TAGS to use the portable_optimized kernels (by default) instead of the reference ones.
+        echo 'TAGS ?= portable_optimized' >> ${TFLITE_MICRO_ROOT}/tools/make/installed_settings.inc
+    fi
     
     # Clean up afterwards because bugs in downlaods from tflite(u)
     # poison VS-code bazel target discovery
